+++ src/main/java/mod/gottsch/forge/gmm/core/entity/monster/IGMMMonster.java
@@ -1,25 +1,28 @@
 package mod.gottsch.forge.gmm.core.entity.monster;
 
+import mod.gottsch.forge.gmm.core.entity.ownership.IOwnable;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.entity.LivingEntity;
 
 import javax.annotation.Nullable;
-import java.util.UUID;
 
 /**
+ * The common contract for all GMM monsters. Extends {@link IOwnable} so every GMM mob carries
+ * ownership (owner UUID + type + lifespan) directly on its own fields/synced data -- the "GMM half"
+ * of the ownership system; non-GMM mobs get the same contract via a capability (see
+ * {@link mod.gottsch.forge.gmm.core.entity.ownership.Ownership}).
+ *
  * @author by Mark Gottschling on 11/2/2025
  */
-public interface IGMMMonster {
+public interface IGMMMonster extends IOwnable {
     // /// NOTE
     // these methods are all from other vanilla interfaces or abstract classes
     // but they are added here to ensure that this singular interface can be used
-    // for all GMM monsters
+    // for all GMM monsters. (getOwnerUUID/setOwnerUUID now come from IOwnable.)
     // ///
     public void defineSynchedData();
     public void addAdditionalSaveData(CompoundTag tag);
     public void readAdditionalSaveData(CompoundTag tag);
-    public UUID getOwnerUUID();
-    void setOwnerUUID(@Nullable UUID uuid);
     // /// END of NOTE
 
     /**